moves hook files from components into hooks folder
